+++ src/util.h
@@ -34,12 +34,12 @@ char *fake_name_quote (const char *c, int quote_percent);
 /* Remove the middle part of the string to fit given length.
  * Use "~" to show where the string was truncated.
  * Return static buffer, no need to free() it. */
-const char *name_trunc (const char *txt, int trunc_len);
+const char *name_trunc (const char *txt, size_t trunc_len);
 
 /* path_trunc() is the same as name_trunc() above but
  * it deletes possible password from path for security
  * reasons. */
-const char *path_trunc (const char *path, int trunc_len);
+const char *path_trunc (const char *path, size_t trunc_len);
 
 /* return a static string representing size, appending "K" or "M" for
  * big sizes.
